requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.3
SQLAlchemy==2.0.31
asyncpg==0.29.0
alembic==1.13.2
celery==5.4.0
redis==5.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
email-validator==2.2.0
httpx==0.27.0
tenacity==8.3.0
psycopg[binary]==3.2.1
python-multipart==0.0.9
Authlib==1.3.1
lumibot==3.18.2